[tint][wgsl] Replace size_t with uint32_t for IDs, Source.
These are used by all AST nodes, and substantially bloat the size of
each and every node. Reducing the size of these substantially reduces
the amount of memory held by a Program, and provides a decent
performance increase due to better cache locality and fewer heap
allocations.

Bug: tint:2129
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/171802
Auto-Submit: Ben Clayton <bclayton@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ src/tint/lang/spirv/validate/validate.cc
@@ -44,8 +44,8 @@ Result<SuccessType> Validate(Slice<const uint32_t> spirv, spv_target_env target_
         [&](spv_message_level_t level, const char*, const spv_position_t& pos, const char* msg) {
             diag::Diagnostic diag;
             diag.message = msg;
-            diag.source.range.begin.line = pos.line + 1;
-            diag.source.range.begin.column = pos.column + 1;
+            diag.source.range.begin.line = static_cast<uint32_t>(pos.line) + 1;
+            diag.source.range.begin.column = static_cast<uint32_t>(pos.column) + 1;
             diag.source.range.end = diag.source.range.begin;
             switch (level) {
                 case SPV_MSG_FATAL:
